+++ src/core/importType.js
@@ -101,10 +101,6 @@ function typeTest(name, context, path) {
   return 'unknown';
 }
 
-export function isScopedModule(name) {
-  return name.indexOf('@') === 0 && !name.startsWith('@/');
-}
-
 export default function resolveImportType(name, context) {
   return typeTest(name, context, resolve(name, context));
 }